Попытался разделить код на модули, но почему-то не работает
