+++ .eslintrc
@@ -8,5 +8,4 @@ env:
   es2017: true
   browser: true
 
-extends:
-  "htmlacademy/vanilla"
+extends: "htmlacademy/vanilla"
